[refactor] kafka - TopicNames 클래스 패키지 이동
- coupon 패키지에서 common 패키지로 이동
